requirements.txt
attrs==23.1.0
beautifulsoup4==4.12.2
bleach==6.1.0
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.1.0
cycler==0.11.0
defusedxml==0.7.1
entrypoints==0.4
fastjsonschema==2.19.0
filelock==3.13.1
fonttools==4.42.1
fsspec==2024.2.0
huggingface-hub==0.20.3
idna==3.6
ipython-genutils==0.2.0
Jinja2==3.0.3
joblib==1.3.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyterlab-pygments==0.2.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.7.2
mistune==0.8.4
mysql-connector==2.2.9
nbclient==0.9.0
nbconvert==5.6.1
nbformat==5.9.2
numpy==1.25.2
packaging==23.1
pandas==2.2.0
pandocfilters==1.5.0
Pillow==10.0.0
platformdirs==4.0.0
Pygments==2.16.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.1
referencing==0.31.0
requests==2.31.0
rpds-py==0.13.0
scikit-learn==1.4.1.post1
scipy==1.12.0
six==1.16.0
skops==0.9.0
soupsieve==2.5
tabulate==0.9.0
testpath==0.6.0
threadpoolctl==3.3.0
tinycss2==1.2.1
tornado==6.3.3
tqdm==4.66.2
traitlets==5.13.0
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.0
webencodings==0.5.1
